shalev change #1 shalev change #2  shalev change #3   shalev change #4

+++ app/src/main/java/com/example/tmpsql/Student.java
@@ -4,13 +4,13 @@ import android.graphics.Bitmap;
 
 public class Student {
     private String firstName,lastName,numClass;
-    private int avgGrade;
+    private double avgGrade;
     Bitmap bitmap;
 
     public Student(String firstName,
                    String lastName,
                    String numClass,
-                   int avgGrade) {
+                   double avgGrade) {
         this.firstName = firstName;
         this.lastName = lastName;
         this.numClass = numClass;
@@ -41,11 +41,11 @@ public class Student {
         this.numClass = numClass;
     }
 
-    public int getAvgGrade() {
+    public double getAvgGrade() {
         return avgGrade;
     }
 
-    public void setAvgGrade(int avgGrade) {
+    public void setAvgGrade(double avgGrade) {
         this.avgGrade = avgGrade;
     }
